+++ src/board/system76/gaze17-3050/include/board/gpio.h
@@ -6,7 +6,6 @@
 #include <ec/gpio.h>
 
 void gpio_init(void);
-void gpio_debug(void);
 
 extern struct Gpio __code ACIN_N;
 extern struct Gpio __code AC_PRESENT;
