+++ src/board/system76/oryp6/gpio.c
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 #include <board/gpio.h>
-#include <common/debug.h>
 #include <common/macro.h>
 
 // clang-format off
@@ -242,40 +241,3 @@ void gpio_init() {
     // SERIRQ
     GPCRM6 = GPIO_ALT;
 }
-
-#if GPIO_DEBUG
-void gpio_debug_bank(
-    char *bank,
-    uint8_t data,
-    uint8_t mirror,
-    uint8_t pot,
-    volatile uint8_t *control
-) {
-    for (char i = 0; i < 8; i++) {
-        DEBUG(
-            "%s%d:\n\tdata %d\n\tmirror %d\n\tpot %d\n\tcontrol %02X\n",
-            bank,
-            i,
-            (data >> i) & 1,
-            (mirror >> i) & 1,
-            (pot >> i) & 1,
-            *(control + i)
-        );
-    }
-}
-
-void gpio_debug(void) {
-#define bank(BANK) gpio_debug_bank(#BANK, GPDR##BANK, GPDMR##BANK, GPOT##BANK, &GPCR##BANK##0)
-    bank(A);
-    bank(B);
-    bank(C);
-    bank(D);
-    bank(E);
-    bank(F);
-    bank(G);
-    bank(H);
-    bank(I);
-    bank(J);
-#undef bank
-}
-#endif
